requirements.txt
streamlit
ultralytics==8.0.196
pillow
numpy
streamlit-webrtc
av